BTUC-18828: Workaround for QTBUG-67376
Maximizing and returning to normal window work as expected when the
windowWillUseStandardFrame:defaultFrame: method returns what it does
originally for frameless windows and for the other windows returning
the newFrame with the X coordinate taken from the QCocoaWindow.

Also reverted the earlier fix since it didn't work with our ToolTips.

Change-Id: I26fdc67deac0c562ff5e0a683b8dcdcbd89c89d7