BTUC-18828: Workaround for QTBUG-67376
Maximizing and returning to normal window work as expected when the
windowWillUseStandardFrame:defaultFrame: method returns what it does
originally for frameless windows and for the other windows returning
the newFrame with the X coordinate taken from the QCocoaWindow.

Also reverted the earlier fix since it didn't work with our ToolTips.

Change-Id: I012f10d3a9570f67b8c18a8b71ae731d037befc0